Menu module

+++ Kafeberus/menu/src/main/java/com/hiberus/models/Product.java
@@ -19,5 +19,5 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private UUID id;
     private String name;
-    private short price;
+    private float price;
 }
